requirements.txt
numpy
pandas
matplotlib
transformers
scipy
torch
lmfit
google-generativeai
python-dotenv
scikit-learn
streamlit
futurehouse-client
edison-client
